refactor: Refactored library structure to reflect object use.
