+++ TildeSummariser/utils/components/base_components.py
@@ -1,0 +1,17 @@
+from abc import ABC, abstractmethod
+
+
+class TextSegmenter(ABC):
+    
+    @abstractmethod
+    def get_segments(self, content)
+    
+class TopicExtractor(ABC):
+
+    @abstractmethod
+    def get_topics(self, segment, n_keywords)
+
+class RedundancyDetector(ABC):
+
+    @abstractmethod
+    def check_entailment(self, text, hypo)
